PhysicsTools

Documentation for PhysicsTools.

PhysicsTools.Utils.CategoricalSetDistribution

PhysicsTools.Utils.apply_rot

PhysicsTools.Utils.fast_linear_interp

PhysicsTools.Utils.fwhm

PhysicsTools.Utils.rand_gamma

PhysicsTools.Utils.repeat_for

PhysicsTools.Utils.sample_cherenkov_track_direction

PhysicsTools.Utils.split_by

PhysicsTools.Utils.split_by!

PhysicsTools.Utils.CategoricalSetDistribution — Type

CategoricalSetDistribution{T, U<:Real}

Represents a Categorical distribution on a set

Examples

p = CategoricalSetDistribution(Set([:EMinus, :EPlus]), Categorical([0.1, 0.9])) rand(p) – returns :EMinus with 10% probability and :Eplus with 90% probability

p = CategoricalSetDistribution(Set([:EMinus, :EPlus]), [0.1, 0.9]) – convenience constructor

PhysicsTools.Utils.apply_rot — Method

apply_rot(a, b, operand)

Calculates rotation matrix obtained by rotating a to b. Apply to operand. Apply the resulting rotation to operand.

PhysicsTools.Utils.fast_linear_interp — Method

fast_linear_interp(x_eval::T, xs::AbstractVector{T}, ys::AbstractVector{T})

Linearly interpolate xs -> ys and evaluate x_eval on interpolation. Assume xs are sorted in ascending order.

PhysicsTools.Utils.fwhm — Method

fwhm(d::UnivariateDistribution, xmode::Real; xlims=(-20, 20))

Calculate FWHM of a univariate distribution

PhysicsTools.Utils.rand_gamma — Method

rand_gamma(shape, scale)

Sample gamma variates when shape > 1

PhysicsTools.Utils.repeat_for — Method

repeat_for(x::AbstractMatrix, n::AbstractVector)

Repeat each slice along the second dimension of x for n times

PhysicsTools.Utils.sample_cherenkov_track_direction — Method

PhysicsTools.Utils.split_by! — Method

split_by!(x::AbstractVector, n::AbstractVector, out)

Write output into out

PhysicsTools.Utils.split_by — Method

split_by(x::AbstractVector, n::AbstractVector)

Split vector x into parts, where the split indices are given by vector n.