PhysicsTools

Documentation for PhysicsTools.

PhysicsTools.Utils.CategoricalSetDistribution

PhysicsTools.Utils.apply_rot

PhysicsTools.Utils.calc_rot_matrix

PhysicsTools.Utils.cart_to_sph

PhysicsTools.Utils.fast_linear_interp

PhysicsTools.Utils.fast_linear_interp

PhysicsTools.Utils.fwhm

PhysicsTools.Utils.integrate_gauss_quad

PhysicsTools.Utils.integrate_gauss_quad

PhysicsTools.Utils.integrate_gauss_quad

PhysicsTools.Utils.rand_gamma

PhysicsTools.Utils.repeat_for

PhysicsTools.Utils.rot_from_ez_fast

PhysicsTools.Utils.rot_to_ez_fast

PhysicsTools.Utils.sample_cherenkov_track_direction

PhysicsTools.Utils.sph_to_cart

PhysicsTools.Utils.split_by

PhysicsTools.Utils.split_by!

PhysicsTools.Utils.transform_integral_range

PhysicsTools.Utils.CategoricalSetDistribution — Type

CategoricalSetDistribution{T, U<:Real}

Represents a Categorical distribution on a set

Examples

p = CategoricalSetDistribution(Set([:EMinus, :EPlus]), Categorical([0.1, 0.9])) rand(p) – returns :EMinus with 10% probability and :Eplus with 90% probability

p = CategoricalSetDistribution(Set([:EMinus, :EPlus]), [0.1, 0.9]) – convenience constructor

PhysicsTools.Utils.apply_rot — Method

apply_rot(a, b, operand)

Calculates rotation matrix obtained by rotating a to b. Applies resulting matrix to operand`.

PhysicsTools.Utils.calc_rot_matrix — Method

calc_rot_matrix(a, b)

Calculates rotation matrix obtained by rotating a to b.

PhysicsTools.Utils.cart_to_sph — Method

cart_to_sph(x::Real, y::Real, z::Real)

Convert cartesian to spherical coordinates. Assumes x, y, z represent a unit vector. Uses ISO convetion (inclination, azimuth).

PhysicsTools.Utils.fast_linear_interp — Method

fast_linear_interp(x_eval::Number, xs::AbstractVector, ys::AbstractVector)

Linearly interpolate xs -> ys and evaluate x_eval on interpolation. Assume xs are sorted in ascending order.

PhysicsTools.Utils.fast_linear_interp — Method

fast_linear_interp(x_eval::Number, knots::AbstractVector, lower::Number, upper::Number)

Linearly interpolate knots and evaluate x_eval on interpolation. Assume knots are equidistant in (lower, upper).

PhysicsTools.Utils.fwhm — Method

fwhm(d::UnivariateDistribution, xmode::Real; xlims=(-20, 20))

Calculate FWHM of a univariate distribution

PhysicsTools.Utils.integrate_gauss_quad — Method

integrate_gauss_quad(f::T, a::Real, b::Real, order::Integer) where {T<:Function}

Integrate f in interval [a, b] using Gauss-Legendre quadrature of order order.

PhysicsTools.Utils.integrate_gauss_quad — Method

integrate_gauss_quad(f::T, a::Real, b::Real) where {T<:Function}

Integrate f in interval [a, b] using 10-point Gauss-Legendre quadrature.

PhysicsTools.Utils.integrate_gauss_quad — Method

integrate_gauss_quad(f::T, a::Real, b::Real, nodes::AbstractVector{U}, weights::AbstractVector{U}) where {T<:Function,U<:Real}

Integrate f in interval [a, b] using Gaussian quadrature with nodes and weights.

PhysicsTools.Utils.rand_gamma — Method

rand_gamma(shape, scale)

Sample gamma variates when shape > 1

PhysicsTools.Utils.repeat_for — Method

repeat_for(x::AbstractMatrix, n::AbstractVector)

Repeat each slice along the second dimension of x for n times

PhysicsTools.Utils.rot_from_ez_fast — Method

rot_to_ez_fast(a::SVector{3,T}, operand::SVector{3,T}) where {T<:Real}

Calc rotation matrix which rotates e_z to a. Applies resulting matrix to operand.

PhysicsTools.Utils.rot_to_ez_fast — Method

rot_to_ez_fast(a::SVector{3,T}, operand::SVector{3,T}) where {T<:Real}

Calc rotation matrix which rotates a to e_z. Applies resulting matrix to operand.

PhysicsTools.Utils.sample_cherenkov_track_direction — Method

sample_cherenkov_track_direction(T::Type)

Sample direction of a Cherenkov track in a particle cascade.

Integrated over an entire particle cascade, this approximates the emission direction of Cherenkov photons. Taken from: https://arxiv.org/abs/1301.5361, p29

PhysicsTools.Utils.sph_to_cart — Method

sph_to_cart(theta::Real, phi::Real)

Convert spherical to cartesian coordinates.

Uses ISO convention (inclination, azimuth).

PhysicsTools.Utils.split_by! — Method

split_by!(x::AbstractVector, n::AbstractVector, out)

Write output into out

PhysicsTools.Utils.split_by — Method

split_by(x::AbstractVector, n::AbstractVector)

Split vector x into parts, where the split indices are given by vector n.

PhysicsTools.Utils.transform_integral_range — Method

transform_integral_range(x::Real, f::T, xrange::Tuple{<:Real,<:Real}) where {T<:Function}

Apply change of interval formula for evaluating quadrature rule on f in interval xrange. See: https://en.wikipedia.org/wiki/Gaussianquadrature#Changeof_interval