PhysicsTools

Documentation for PhysicsTools.

PhysicsTools.Types.Cascade

PhysicsTools.Types.PEMinus

PhysicsTools.Types.PEPlus

PhysicsTools.Types.PGamma

PhysicsTools.Types.PHadronShower

PhysicsTools.Types.PLightSabre

PhysicsTools.Types.PMuMinus

PhysicsTools.Types.PMuPlus

PhysicsTools.Types.PNuE

PhysicsTools.Types.PNuEBar

PhysicsTools.Types.PNuMu

PhysicsTools.Types.PNuMuBar

PhysicsTools.Types.PNuTau

PhysicsTools.Types.PNuTauBar

PhysicsTools.Types.PUnknown

PhysicsTools.Types.Particle

PhysicsTools.Types.Track

PhysicsTools.Types.is_neutrino

PhysicsTools.Types.particle_shape

PhysicsTools.Types.pdg_code

PhysicsTools.Types.ptype_for_code

PhysicsTools.Types.Cascade — Type

Subtype of ParticleShape, used to classify signatures in detector. PEPlus, PEMinus, PGamma and PHadronShower have the shape Cascade; while PMuMinus, PMuPlus and PLightSabre have the shape Track. particle_shape() can be used to return shape of particle signature.

PhysicsTools.Types.PEMinus — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PEPlus — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PGamma — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PHadronShower — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PLightSabre — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PMuMinus — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PMuPlus — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuE — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuEBar — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuMu — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuMuBar — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuTau — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PNuTauBar — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.PUnknown — Type

Subtype of ParticleType, used to classify particles. pdg_code() can be used to find the corresponding particle PDF code.

PhysicsTools.Types.Particle — Type

Particle{T,PType<:ParticleType}

Struct containing information of a single particle

Fields

position::SVector{3,T}: Particle position given as a three dimensional static vector in detector coordinates

direction::SVector{3,T}: Particle direction of motion given as a three dimensional static vector in detector coordinates

time::T: Time of the event this particle corresponds to, given in unites of ns

energy::T: Total energy of the particle given in unites of GeV

length::T: Length of particle track given in unites of m

type::Type{PType}: Particle type, chosen from any ParticleType subtype (PEPlus, PEMinus, PGamma, PMuPlus, PMuMinus, PNuE, PNuMu, PNuTau, PNuEBar, PNuMuBar, PNuTauBar, PHadronShower, PLightSabre, PUnknown)

PhysicsTools.Types.Track — Type

Subtype of ParticleShape, used to classify signatures in detector. PEPlus, PEMinus, PGamma and PHadronShower have the shape Cascade; while PMuMinus, PMuPlus and PLightSabre have the shape Track. particle_shape() can be used to return shape of particle signature.

PhysicsTools.Types.is_neutrino — Method

is_neutrino(particle)

Return true if particle is of any neutrino subtype (PNuE, PNuEBar, PNuMu, PNuMuBar, PNuTau, PNuTauBar), otherwise return false.

Arguments

particle::Type: Particle to check wheather is a neutrino.

PhysicsTools.Types.particle_shape — Method

particle_shape(partricle)

Return struct corresponding to shape of argument-particle's signature

Arguments

particle::Type: Particle to check shape of.

PhysicsTools.Types.pdg_code — Method

pdg_code(particle)

Return integer corresponding to the PDG code of a givern particle. ptypeforcode() can be used to find the struct of the corresponding particle.

Arguments

particle::Type: Particle to return PDG code of.

PhysicsTools.Types.ptype_for_code — Method

ptype_for_code(code::Integer)

Return particle struct corresponding to a given PDG code. pdg_code() can be used to find the PDG code of a corresponding particle.

Arguments

code::Integer: PDG code to get particle struct of.